requirements.txt
numpy==1.23.5
Pillow==9.5.0
requests==2.25.1
# pymongo==3.11.3
Flask==2.3.3
pytest==8.3.4
psycopg2-binary==2.9.10 
sqlalchemy==2.0.36
python-dotenv==1.0.1